Make tests succeed reliably.

+++ src/auth/session/storage/__tests__/redis.test.ts
@@ -1,13 +1,15 @@
 import * as child_process from 'child_process';
 import {promisify} from 'util';
 import {resolve} from 'path';
+import * as net from 'net';
 
-import {createClient} from 'redis';
+// import {createClient, RedisClientType} from 'redis';
 
 import {RedisSessionStorage} from '../redis';
 
 import {batteryOfTests} from './battery-of-tests';
-import {poll} from './utils';
+// import {poll, wait} from './utils';
+import {poll, connectSocket, waitForData} from './utils';
 
 const exec = promisify(child_process.exec);
 
@@ -17,8 +19,8 @@ const dbURL = new URL('redis://shopify:passify@localhost/1');
 jest.setTimeout(20000);
 
 describe('RedisSessionStorage', () => {
-  let storage: RedisSessionStorage;
-  let containerId: string;
+  let storage: RedisSessionStorage | undefined;
+  let containerId: string | undefined;
   beforeAll(async () => {
     const configPath = resolve(__dirname, './redis.conf');
     const runCommand = await exec(
@@ -30,11 +32,18 @@ describe('RedisSessionStorage', () => {
     await poll(
       async () => {
         try {
-          const client = createClient({
-            url: dbURL.toString(),
-          });
-          await client.connect();
-          await client.quit();
+          const socket = new net.Socket();
+          await Promise.race([
+            connectSocket(socket, {
+              port: 6379,
+              host: 'localhost',
+            }),
+          ]);
+          // Invalid command will trigger a text response from the server,
+          // indicating that the server is ready and listening.
+          socket.end('HELP I’M TRAPPED IN A REDIS FACTORY\n');
+          const data = await waitForData(socket);
+          if (data === 'no data') return false;
         } catch {
           return false;
         }
@@ -47,9 +56,9 @@ describe('RedisSessionStorage', () => {
   });
 
   afterAll(async () => {
-    await storage.disconnect();
-    await exec(`docker rm -f ${containerId}`);
+    await storage?.disconnect();
+    if (containerId) await exec(`docker rm -f ${containerId}`);
   });
 
-  batteryOfTests(async () => storage);
+  batteryOfTests(async () => storage!);
 });
